+++ NEWS.md
@@ -1,6 +1,10 @@
-# councilR v0.1.0-beta.1  
+# v0.1.0.9000 (2020-10-21)  
 
-2020-05-26
+## New functions
+
+* `import_from_gpkg()` function allows the user to import an `sf` object from a geopackage URL. This is particularly useful for accessing Minnesota Geospatial Commons if you are unable to access the Council's ArcGIS catalog. 
+
+# v0.1.0  (2020-05-26)
 
 ## New functions  
 
@@ -9,4 +13,4 @@
 ## Housekeeping  
 
 * Added a `NEWS.md` file to track changes to the package.  
-* Updataed [Contributor Code of Conduct](.github/CODE_OF_CONDCT.md).  
+* Updated [Contributor Code of Conduct](.github/CODE_OF_CONDCT.md).